--- cea/examples/tooloftoday.py
@@ -1,57 +1,0 @@
-"""
-This tool takes the occupancy.dbf file and computes the product of that file with an X variable parameter stored in the config.
-The results are saved in the output folder of CEA as a CSV file.
-"""
-from __future__ import division
-from __future__ import print_function
-
-import os
-import cea.config
-import cea.inputlocator
-from cea.utilities.dbf import dbf_to_dataframe
-import pandas as pd
-
-
-__author__ = "Luis Santos"
-__copyright__ = "Copyright 2018, Architecture and Building Systems - ETH Zurich"
-__credits__ = ["Luis Santos, Jimeno Fonseca"]
-__license__ = "MIT"
-__version__ = "0.1"
-__maintainer__ = "Daren Thomas"
-__email__ = "cea@arch.ethz.ch"
-__status__ = "Production"
-
-
-def tooloftoday_main(locator, config):
-    """this is where the action happens if it is more than a few lines in ``main``.
-    NOTE: ADD YOUR SCRIPT'S DOCUMENATION HERE (how)
-    NOTE: RENAME THIS FUNCTION (SHOULD PROBABLY BE THE SAME NAME AS THE MODULE)
-    """
-
-    occ_path = locator.get_building_occupancy()
-    dataframe_occupancy = dbf_to_dataframe(occ_path)
-    multiplier = config.tooloftoday.multiplier
-    total = dataframe_occupancy*int(multiplier)
-    total.to_csv(locator.get_totaloccupancy())
-    pass
-
-
-def main(config):
-    """
-    This is the main entry point to your script. Any parameters used by your script must be present in the ``config``
-    parameter. The CLI will call this ``main`` function passing in a ``config`` object after adjusting the configuration
-    to reflect parameters passed on the command line - this is how the ArcGIS interface interacts with the scripts
-    BTW.
-
-    :param config:
-    :type config: cea.config.Configuration
-    :return:
-    """
-    assert os.path.exists(config.scenario), 'Scenario not found: %s' % config.scenario
-    locator = cea.inputlocator.InputLocator(config.scenario)
-
-    tooloftoday_main(locator, config)
-
-
-if __name__ == '__main__':
-    main(cea.config.Configuration())
